requirements.txt
# Space Pong - Dependencies
# Core game engine
pygame==2.5.2

# Image processing for sprite extraction
pillow==10.2.0

# Optional: For optimized physics calculations
numpy==1.26.3
